Styleguide, more unit tests, and custom filter key functionality for checkboxFilterGroup directive
