+++ docs/index.html
@@ -21,7 +21,7 @@
             <li class="divider-vertical"></li>
             <li>
                 <p class="navbar-text">
-                    API Docs for Version: <b>1.0.1</b>
+                    API Docs for Version: <b>1.1.1</b>
                 </p>
             </li>
         </div>
