+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "c2fo-ui-components",
-  "version": "1.0.1",
+  "version": "1.1.1",
   "description": "C2FO core ui components",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "grunt build"
   },
