+++ src/index.js
@@ -1,8 +1,5 @@
 "use strict";
 
-var core = require("./core"),
-    modules =
-
 angular.module("c2fo.ui", [
     require("./core"),
     require("./modules")
