+++ src/modules/index.js
@@ -3,8 +3,8 @@
 var moduleName = "c2fo.ui.modules";
 
 angular.module(moduleName, [
-    require("./checkboxFilterGroup/checkboxFilterGroup.js"),
-    require("./pager/pager.js")
+    require("./checkboxFilterGroup"),
+    require("./pager")
 ]);
 
 module.exports = moduleName;
